Refactor enum cases to lowercase
Changes have been made to the enum cases in EventTransparency.php, Event.php, and EventClassification.php, converting them from uppercase to lowercase. This has been reflected also in the tests in EventTest.php. The revisions are aimed at bolstering adherence to the common coding standard of defining constants in lowercase.